Data.Expire

Documentation

expireSteps :: Int

expireSteps is a global number of steps taken for expiration. The idea is that the monad will throw some sort of exception, or result in Nothing, when more than this many Functor, Applicative, or Monad steps (actions) have been taken.

newtype Expire a

Pure expiration, return Nothing to expire

Constructors

Expire

Fields

getExpire :: (Int, Maybe a)

newtype ExpireIO a

Expire with IO, use a Weak reference outside of the Maybe that Expire uses.

Hopefully, ExpireIO will give near-instant expiration, not like Expire, which seems to take around 1ns per action after expiration.

Constructors

ExpireIO

Fields

runExpireIO :: IO (Weak (IORef (Int, Maybe a)))

stepExpire :: Int -> a -> (Int, Maybe a)

Note: stepExpire will never halt if Int is negative, so instead it throws an error

stepExpireMaybe :: Int -> Maybe a -> (Int, Maybe a)

stepExpire for Maybe

coerceWeak :: Weak a -> Weak b

Note: This is only safe if a Weak pointer has _already_ been dereferenced

getExpireIO :: ExpireIO a -> IO (Maybe a)

Deref an ExpireIO, only returning the value, or return Nothing

dumpExpireIO :: ExpireIO a -> IO (Maybe (Int, a))

Dereference an ExpireIO, or return Nothing

killExpireIO :: ExpireIO a -> IO ()

finalize an ExpireIO

newtype ExpireE a

Expire that throws and error instead of returning Nothing

Constructors

ExpireE

Fields

getExpireE :: (Int, a)