Data.Timers.Example

Documentation

data Map k a :: * -> * -> *

A Map from keys k to values a.

gapless :: (Bits t, Num t) => t -> Bool

Mystery function... (read the source)

gapless' :: (Bits t, Num t) => t -> Bool

Helper function to gapless

gaplessEq :: (Bits a, Num a) => a -> Bool

Are gapless and gapless' equal?

countGapless :: (Bits a, Enum a, Num a, Num b) => a -> b

Number of gapless numbers in [1..n]

countGapless' :: (Bits a, Enum a, Num a, Num b) => a -> b

Number of gapless' numbers in [1..n]

nextGapless :: (Bits a, Num a) => a -> a

What it says on the box

toMathematica :: Show a => [[a]] -> String

Convert a list of lists to a list of Mathematica lists.

onlyGapless :: (Bits t, Num t) => [t] -> [t]

One-liner to export gapless to mathematica:

 concatMap (x->concat["{",show(head x),",",show(length x),"},"]).group . (x->zipWith(-)(tail x)x) . onlyGapless $ [1..100]

See only_gapless.md

onlyGaplessMathematica :: (Bits t, Num t, Show t) => [t] -> String

onlyGapless, differences, grouped, then converted to Mathematica lists. {--}

differences :: Num a => [a] -> [a]

First differences:

 [x0, x1, x2, x3..] -> [x1 - x0, x2 - x1, x3 - x2..]

differences_ :: Num a => [a] -> [a]

Should be equivalent to differences

prop_differences_ :: [Int] -> Bool

Compare differences and differences_, specialized for testing

getRuns :: (Num a, Num b, Eq a) => [a] -> [b]

Warning, this function is incomplete.

λ> take 10 . getRuns . repeat $ 0 == _|_
λ> take 10 . getRuns . repeat $ 1 == _|_
λ> take 10 . getRuns . cycle $ [0] == _|_
λ> take 10 . getRuns . cycle $ [1] == _|_
λ> take 10 . getRuns . cycle $ [0,0] == _|_
λ> take 10 . getRuns . cycle $ [0,1] == [1,1,1,1,1,1,1,1,1,1]
λ> take 10 . getRuns . cycle $ [1,0] == [1,1,1,1,1,1,1,1,1,1]
λ> take 10 . getRuns . cycle $ [1,1] == _|_
λ> take 10 . getRuns . cycle $ [0,0,0] == _|_
λ> take 10 . getRuns . cycle $ [0,0,1] == [2,2,2,2,2,2,2,2,2,2]
λ> take 10 . getRuns . cycle $ [0,1,0] == [1,2,2,2,2,2,2,2,2,2]
λ> take 10 . getRuns . cycle $ [0,1,1] == [1,1,1,1,1,1,1,1,1,1]
λ> take 10 . getRuns . cycle $ [1,0,0] == [2,2,2,2,2,2,2,2,2,2]
λ> take 10 . getRuns . cycle $ [1,0,1] == [1,1,1,1,1,1,1,1,1,1]
λ> take 10 . getRuns . cycle $ [1,1,0] == [1,1,1,1,1,1,1,1,1,1]
λ> take 10 . getRuns . cycle $ [1,1,1]

It's a fun autamata, note that it must be passed an infinite list, or else it'll hit an exception. Even then, on some infinite lists, it'll loop forever without printing (if it never hits 1's).

getRunsA :: (Num a, Num b, Eq a) => [a] -> [b]

A "match [] first" version of getRuns

getRunsB :: (Num a, Num b, Eq a) => [a] -> [b]

A "match [] last" version of getRuns

getRuns' :: (Num a, Num b, Eq a) => b -> [a] -> [b]

Warning, this function is incomplete.

getRuns'A :: (Num a, Num b, Eq a) => b -> [a] -> [b]

A "match [] first" version of getRuns'

getRuns'B :: (Num a, Num b, Eq a) => b -> [a] -> [b]

A "match [] last" version of getRuns'

addPosition

Arguments

:: (Ord k0, Eq a, Num a)

=> (a, k0)
(position, value)

-> Map k0 ((a, a, a, a, a), [(a, a, a, a, a)])
Map Value (last input position, list of its previous runs)

-> Map k0 ((a, a, a, a, a), [(a, a, a, a, a)])
Map Value (new input position, list of its previous runs)

Add a position to the Map

type Vec2 = ContVec 2

Abbrev. to make it easier to make the code generic

type Vec4 = ContVec 4

Abbrev. to make it easier to make the code generic

type Vec5 = ContVec 5

Abbrev. to make it easier to make the code generic

addPosition' :: (Ord k0, Eq a, Num a) => a -> k0 -> Map k0 (Vec5 a, [Vec5 a]) -> Map k0 (Vec5 a, [Vec5 a])

A version of addPosition using Vec5 instead of tuples

theStream :: (Enum a, Enum b, Num b, Bits b, Num a) => ([(a, b)], Map k v)

The initial state of the stream

printMap :: (Show b, Show c) => (a1, Map b (a, [(c, c, c, c, c)])) -> IO ()

Print the resulting Map

printMap' :: (Show a0, Show a2) => (a, Map a0 (a1, [Vec5 a2])) -> IO ()

Print the resulting Map, specialized to Vec5's instead of tuples

printEveryNOfStream :: (Eq a, Bits k0, Num k0, Num a, Enum k0, Enum a, Ord k0, Show k0, Show a) => Int -> IO ([(a, k0)], Map k0 ((a, a, a, a, a), [(a, a, a, a, a)]))

Print every nth value of the stream, specialized to the following value (it's an infinite loop, so the return value simply encodes which types the algorithm is specialized to).

printEveryNOfStream' :: (Eq a, Bits k0, Num k0, Num a, Enum k0, Enum a, Ord k0, Show (ContVec 5 a), Show k0, Show a) => Int -> IO ([(a, k0)], Map k0 (Vec5 a, [Vec5 a]))

printEveryNOfStream, secialized to ContVec's and using addPosition'

returnEveryN

Arguments

:: (Eq c, Show b, Show c, Ord b, Enum c, Enum b, Num c, Num b, Bits b)

=> Int
How often

-> IO ([(c, b)], Map b ((c, c, c, c, c), [(c, c, c, c, c)]))
Results from theStream

See printEveryNOfStream, except is uses sideEffectEveryN instead of printEveryN

returnEveryN'

Arguments

:: (Eq a2, Enum a2, Num a2, Bits a1, Enum a1, Num a1, Ord a1, Show a2, Show a1)

=> Int
How often

-> IO ([(a2, a1)], Map a1 (Vec5 a2, [ContVec 5 a2]))
Results from theStream

A version of returnEveryN using addPosition' and printMap'